pyproject.toml
[tool.poetry]
name = "responsive-gen"
version = "0.1.0"
description = "Responsive webpage generation from multi-viewport wireframe sketches using LLMs"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "responsive_gen"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-anthropic = "^0.1.0"
langchain-core = "^0.1.0"
openai = "^1.10.0"
anthropic = "^0.18.0"
streamlit = "^1.31.0"
pillow = "^10.2.0"
numpy = "^1.26.0"
opencv-python = "^4.9.0"
playwright = "^1.41.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
transformers = "^4.37.0"
torch = "^2.2.0"
torchvision = "^0.17.0"
clip-by-openai = "^1.0"
scipy = "^1.12.0"
scikit-image = "^0.22.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pandas = "^2.2.0"
jsonschema = "^4.21.0"
tenacity = "^8.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
ipython = "^8.20.0"
jupyter = "^1.0.0"

[tool.poetry.scripts]
responsive-gen = "cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

